Extracted the logic to a different file.
