+++ demos/stock-timeseries/demo.html
@@ -21,6 +21,6 @@
             <input type="file" id="postman-json" accept=".json,application/json" />
         </p>
         <div id="container"></div>
-        <script src="./demo.js"></script>
+        <script type="module" src="./demo.js"></script>
     </body>
 </html>
